Improved export and added import.

+++ chuachua/src/main/java/org/byeautumn/chuachua/io/IOUntil.java
@@ -1,12 +1,15 @@
 package org.byeautumn.chuachua.io;
 
 import org.bukkit.Bukkit;
+import org.bukkit.Material;
+import org.bukkit.World;
+import org.bukkit.block.Block;
+import org.bukkit.util.Vector;
 import org.byeautumn.chuachua.Universe;
 
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.io.*;
+import java.util.ArrayList;
+import java.util.List;
 
 public class IOUntil {
     public static final String IO_DIR_NAME = "io";
@@ -15,27 +18,35 @@ public class IOUntil {
     public static boolean checkDirExistAndCreateItIfNot(File dir) {
         if (!dir.exists()) {
             System.out.println("IO folder doesn't exist. Trying to create it ...");
-            try {
-                dir.createNewFile();
-            } catch (IOException ioe) {
-                System.err.println("IO folder creation failed: " + ioe.getMessage());
-                return false;
+            if (dir.mkdirs()) {
+                System.err.println("Failed to create " + dir.getName());
             }
-            System.out.println("IO folder is created successfully.");
+
+            System.out.println("The folder " + dir.getName() + "is created successfully.");
         }
 
         return true;
     }
 
-    public static boolean saveExportIntoAIOFile(String chunkName, String exportContent) {
+    public static File getIODir() {
         File serverDir = Bukkit.getWorldContainer();
-        File ioDir = new File(serverDir, IO_DIR_NAME);
+        return new File(serverDir, IO_DIR_NAME);
+    }
+
+    public static boolean saveExportIntoAIOFile(String chunkName, String exportContent) {
+        File ioDir = getIODir();
         if (!IOUntil.checkDirExistAndCreateItIfNot(ioDir)) {
             System.err.println("The IO folder doesn't exist and failed on trying to create it.");
             return false;
         }
 
-        String fileName = getAbsoluteFilePath(ioDir, chunkName + "." + IO_EXTENSION);
+        final String fileName = getAbsoluteFilePath(ioDir, chunkName + "." + IO_EXTENSION);
+        final File file = new File(fileName);
+        if (file.exists()) {
+            System.err.println("'" + chunkName + "' exists already. Please give the chunk exportChunk another name.");
+            return false;
+        }
+
         try (BufferedWriter writer = new BufferedWriter(new FileWriter(fileName))) {
             writer.write(exportContent);
             System.out.println("File created and saved successfully: " + fileName);
